+++ src/Pages/ManageAllOrder/ManageAllOrder.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { useState } from 'react';
 import { useEffect } from 'react';
 import { Container, Table } from 'react-bootstrap';
+import { Link } from 'react-router-dom';
 import "./ManageAllOrder.css"
 
 const ManageAllOrder = () => {
@@ -11,25 +12,25 @@ const ManageAllOrder = () => {
         fetch('http://localhost:5000/manageOrder')
             .then(res => res.json())
             .then(data => setUsers(data))
-    }, [isDelete])
-    const handelUpdate = () => {
-
-    }
+    }, [isDelete]);
 
     const handelDelete = (id) => {
-        fetch(`http://localhost:5000/deleteOrders/${id}`, {
-            method: 'DELETE'
-        })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount) {
-                    alert('Are you sure to Delete')
-                    setIsDelete(true)
-                }
-                else {
-                    setIsDelete(false)
-                }
+        const proceed = window.confirm('Are you Sure to delete')
+        if (proceed) {
+            fetch(`http://localhost:5000/deleteOrders/${id}`, {
+                method: 'DELETE'
             })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.deletedCount) {
+                        alert('Delete This Order')
+                        setIsDelete(true)
+                    }
+                    else {
+                        setIsDelete(false)
+                    }
+                });
+        }
     }
     return (
         <div className='my-4'>
@@ -41,6 +42,7 @@ const ManageAllOrder = () => {
                             <th>Email</th>
                             <th>Order Name</th>
                             <th>Package Price</th>
+                            <th>Order Status</th>
                             <th>Update</th>
                             <th>Delete</th>
                         </tr>
@@ -51,7 +53,8 @@ const ManageAllOrder = () => {
                                 <td>{user.email}</td>
                                 <td>{user.tourName}</td>
                                 <td>{user.price}</td>
-                                <td><i class="fas fa-edit update-icon" onClick={handelUpdate}></i></td>
+                                <td>{user.status}</td>
+                                <td><Link to={`/updateStatus/${user._id}`}><i class="fas fa-edit update-icon"></i></Link></td>
                                 <td><i class="fas fa-trash-alt delete-icon" onClick={() => handelDelete(user._id)}></i></td>
                             </tr>
                             )
